requirements.txt
aiofiles
aiohttp
apscheduler
asyncio
uvloop
bs4
ffmpeg-python
future==1.0.0
gitpython
heroku3
httpx[http2]==0.28.1
motor 
pillow==11.2.1
psutil
pytgfork
python-dotenv
ntgcalls==2.0.6
py-tgcalls==2.2.8
pyyaml==6.0.2
requests
speedtest-cli
spotipy
tgcrypto
unidecode
yt-dlp==2025.10.22
git+https://github.com/BillaSpace/youtube-search-python.git
